pyproject.toml
[build-system]
requires = ["hatchling>=1.21.1"]
build-backend = "hatchling.build"

[project]
name = "catalog-enrichment"
version = "0.1.0"
description = "A system for enriching product catalog data with additional metadata, descriptions, and categorization."
requires-python = ">=3.11"
dependencies = [
  "uvicorn[standard]==0.38.0",
  "starlette==0.49.3",
  "fastapi==0.121.0",
  "python-multipart==0.0.20",
  "openai==2.7.1",
  "python-dotenv==1.2.1",
  "requests==2.32.5",
  "httpx==0.28.1",
  "pillow==12.0.0",
  "pyyaml==6.0.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

[tool.uv]
dev-dependencies = [
  "pytest==8.4.2",
  "pytest-mock==3.14.0",
  "pytest-asyncio==0.25.2",
  "httpx==0.28.1",
]


